+++ composenav/src/main/java/com/adamkobus/compose/navigation/NavActionConsumer.kt
@@ -1,7 +1,7 @@
 package com.adamkobus.compose.navigation
 
-import com.adamkobus.compose.navigation.data.NavAction
-import com.adamkobus.compose.navigation.data.NavActionWrapper
+import com.adamkobus.compose.navigation.action.NavAction
+import com.adamkobus.compose.navigation.action.NavActionWrapper
 
 interface NavActionConsumer {
     fun offer(action: NavAction)