+++ composenav/src/main/java/com/adamkobus/compose/navigation/NavActionVerifier.kt
@@ -1,10 +1,11 @@
 package com.adamkobus.compose.navigation
 
-import com.adamkobus.compose.navigation.data.INavDestination
-import com.adamkobus.compose.navigation.data.NavAction
+import com.adamkobus.compose.navigation.action.NavAction
+import com.adamkobus.compose.navigation.action.NavigateAction
+import com.adamkobus.compose.navigation.destination.INavDestination
 
 /**
- * Purpose of this class is to verify if [NavAction] execution is allowed based on the application state.
+ * Purpose of this class is to verify if [NavigateAction] execution is allowed based on the application state.
  *
  * As an example, it could be used to prevent duplicate nav action if user clicks list item twice.
  *
@@ -16,9 +17,13 @@ interface NavActionVerifier {
      * @param currentDestination Represents the destination displayed to the user at the moment of performing this check
      * @param action Action that was produced via [NavigationState.postNavAction]
      *
-     * @return false if [action] should be discarded.
-     * True should be returned if [action] can be processed further for given [currentDestination]
-     * or if [action] is not recognized by this [NavActionVerifier] implementation.
+     * @return [VerifyResult.Discard] if [action] should be discarded.
+     * [VerifyResult.Allow] should be returned if this [NavActionVerifier] doesn't want to block the [action]
      */
-    fun isNavActionAllowed(currentDestination: INavDestination, action: NavAction): Boolean
+    fun isNavActionAllowed(currentDestination: INavDestination, action: NavAction): VerifyResult
+}
+
+enum class VerifyResult {
+    Allow,
+    Discard
 }