+++ composenav/src/main/java/com/adamkobus/compose/navigation/data/NavActionWrapper.kt
@@ -1,5 +1,0 @@
-package com.adamkobus.compose.navigation.data
-
-interface NavActionWrapper {
-    val action: NavAction
-}