+++ composenav/src/main/java/com/adamkobus/compose/navigation/ui/NavComposable.kt
@@ -5,17 +5,29 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
+import com.adamkobus.compose.navigation.data.NavGraph
 
 /**
- * This components acts as a bridge between the view layer that operates with [NavHostController] and model.
- */
+ * This components acts as a bridge between the owner of [navController] and the navigation state model.
+ *
+ * @param observedGraphs If not empty, then only NavActions that originate from the graphs included in this list will
+ * be processed. All NavActions will be processed if [observedGraphs] is empty
+ **/
 @Composable
-fun NavComposable(navController: NavHostController) {
+fun NavComposable(
+    navController: NavHostController,
+    vararg observedGraphs: NavGraph
+) {
     val vm = hiltViewModel<NavComposableVM>()
     LifecycleAwareComponent(vm)
-    val pendingActions = vm.pendingActions.value
-    LaunchedEffect(key1 = pendingActions) {
-        vm.processNavActions(navController = navController)
+    vm.graphsParam.bind(observedGraphs.toList())
+
+    val pendingAction = vm.pendingActionState.value
+    LaunchedEffect(key1 = pendingAction) {
+        if (pendingAction is PendingActionState.Present) {
+            pendingAction.action.navigate(navController)
+            pendingAction.complete()
+        }
     }
     val currentBackStackEntry = navController.currentBackStackEntryFlow.collectAsState(initial = null)
     LaunchedEffect(key1 = currentBackStackEntry.value) {