+++ demo-settings/src/main/java/com/adamkobus/compose/navigation/demo/settings/nav/SettingsGraph.kt
@@ -4,21 +4,20 @@ import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.navigation.NavGraphBuilder
 import com.adamkobus.compose.navigation.data.NavGraph
 import com.adamkobus.compose.navigation.demo.settings.ui.home.SettingsHomeScreen
+import com.adamkobus.compose.navigation.destination.NavDestination
 import com.adamkobus.compose.navigation.ext.composableDestination
 import com.adamkobus.compose.navigation.ext.composableNavigation
 
-object SettingsGraph : NavGraph {
-    override val name: String = "settingsGraph"
+object SettingsGraph : NavGraph("settingsGraph") {
+
+    override fun startDestination(): NavDestination = SettingsHome
 
     val SettingsHome = navDestination("home")
 }
 
 @OptIn(ExperimentalAnimationApi::class)
 fun NavGraphBuilder.settingsGraph() {
-    composableNavigation(
-        graph = SettingsGraph,
-        startDestination = SettingsGraph.SettingsHome
-    ) {
+    composableNavigation(SettingsGraph) {
 
         composableDestination(SettingsGraph.SettingsHome) {
             SettingsHomeScreen()