+++ demo/src/main/java/com/adamkobus/compose/navigation/demo/nav/Actions.kt
@@ -1,11 +1,11 @@
 package com.adamkobus.compose.navigation.demo.nav
 
-import com.adamkobus.compose.navigation.data.GlobalDestination
-import com.adamkobus.compose.navigation.data.NavAction
-import com.adamkobus.compose.navigation.data.PopBackStackDestination
+import com.adamkobus.compose.navigation.action.NavAction
+import com.adamkobus.compose.navigation.action.NavActionWrapper
+import com.adamkobus.compose.navigation.data.GlobalGraph
 import com.adamkobus.compose.navigation.demo.settings.nav.SettingsGraph
 
-sealed class FromSplash(action: NavAction) : NavAction(action) {
+sealed class FromSplash(action: NavAction) : NavActionWrapper(action) {
     object ToWelcome : FromSplash(
         AppGraph.SplashScreen to AppGraph.WelcomeScreen navigate {
             popUpTo(AppGraph.name)
@@ -14,7 +14,7 @@ sealed class FromSplash(action: NavAction) : NavAction(action) {
     )
 }
 
-sealed class FromWelcome(action: NavAction) : NavAction(action) {
+sealed class FromWelcome(action: NavAction) : NavActionWrapper(action) {
     object ToCatsList : FromWelcome(
         AppGraph.WelcomeScreen to AppGraph.CatsList navigate {
             popUpTo(AppGraph.name)
@@ -30,22 +30,30 @@ sealed class FromWelcome(action: NavAction) : NavAction(action) {
     )
 }
 
-sealed class FromCatsList(action: NavAction) : NavAction(action) {
+sealed class FromCatsList(action: NavAction) : NavActionWrapper(action) {
     // navigate is optional. It's needed only if you want to provide custom options to the NavOptionsBuilder
     // If you're fine with default navigation behaviour, then it's ok to skip "navigate" block
     class ToCatDetails(catId: Int) : FromCatsList(AppGraph.CatsList to AppGraph.CatDetails arg catId)
 }
 
-sealed class FromDogsList(action: NavAction) : NavAction(action) {
+sealed class FromCatDetails(action: NavAction) : NavActionWrapper(action) {
+    object Back : FromCatDetails(AppGraph.CatDetails to AppGraph.Back)
+}
+
+sealed class FromDogsList(action: NavAction) : NavActionWrapper(action) {
     class ToDogDetails(dogId: Int) : FromDogsList(AppGraph.DogsList to AppGraph.DogDetails arg dogId)
 }
 
-sealed class FromGlobal(action: NavAction) : NavAction(action) {
-    object ToSettings : FromGlobal(GlobalDestination to SettingsGraph)
-
-    object ToDemoDialog : FromGlobal(GlobalDestination to AppGraph.DemoDialog)
+sealed class FromDogDetails(action: NavAction) : NavActionWrapper(action) {
+    object Back : FromDogDetails(AppGraph.DogDetails to AppGraph.Back)
 }
 
-sealed class FromDemoDialog(action: NavAction) : NavAction(action) {
-    object Dismiss : FromDemoDialog(NavAction(AppGraph.DemoDialog, PopBackStackDestination, navigateWithController = { it.popBackStack() }))
+sealed class FromGlobal(action: NavAction) : NavActionWrapper(action) {
+    object ToSettings : FromGlobal(GlobalGraph to SettingsGraph)
+
+    object ToDemoDialog : FromGlobal(GlobalGraph to AppGraph.DemoDialog)
+}
+
+sealed class FromDemoDialog(action: NavAction) : NavActionWrapper(action) {
+    object Dismiss : FromDemoDialog(AppGraph.DemoDialog to AppGraph.Back)
 }